gcs: Split gcsv2 handlers
